Code reorganization, base eggification, system executable
