+++ bitkey/algo.py
