+++ bitkey/mnemonic.py
