+++ bitkey/transport_serial.py
