--- protobuf/build.sh
@@ -1,5 +1,0 @@
-#!/bin/bash
-
-cd `dirname $0`
-
-protoc --python_out=../src/ *.proto
