--- tests/run.sh
@@ -1,2 +1,0 @@
-#!/bin/bash
-python -m unittest discover
